requirements.txt
python-telegram-bot==22.5
environs==14.5.0